Expand improvements

+++ src/integer.h
@@ -18,7 +18,8 @@ public:
     virtual bool __eq__(const Basic &o) const;
     virtual std::string __str__() const;
 
-    int as_int(); // Convert to "int", raise an exception if it is too large
+    // Convert to "int", raise an exception if it does not fit
+    signed long int as_int();
     inline bool is_zero() { return this->i == 0; }
     inline bool is_one() { return this->i == 1; }
 };
